# openapi.yaml
openapi: "3.1.0"
info:
  title: Agent Kanban REST API
  description: REST API for managing Kanban projects, tickets, columns, workflows, roles, agent communications (conversations), and global settings. Each MCP server process is associated with a role that determines permissions.
  version: "1.0.0"
  contact:
    name: Agent Kanban Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
paths:
...
  /projects/{slug}/tickets:
    get:
      summary: List tickets with filters
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: column
          in: query
          description: Filter by column slug
          required: false
          schema:
            type: string
        - name: priority
          in: query
          description: Filter by priority (1=urgent, 2=high, 3=medium, 4=low, 5=trivial)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: labels
          in: query
          description: Comma-separated labels filter (any match)
          required: false
          schema:
            type: string
        - name: parent_id
          in: query
          description: Filter by parent ticket ID (returns sub-tickets)
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: "Items per page (default: 20, max: 1000)"
          required: false
          schema:
            type: integer
            default: 20
            maximum: 1000
        - name: sort_by
          in: query
          description: "Sort field (default: created_at)"
          required: false
          schema:
            type: string
            enum:
              - priority
              - created_at
              - updated_at
              - title
            default: created_at
        - name: sort_order
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - name: mode
          in: query
          description: "Filter mode. 'not-blocked': unblocked only. 'top-level-tickets': parent tickets only"
          required: false
          schema:
            type: string
            enum:
              - not-blocked
              - top-level-tickets
        - name: all_tickets
          in: query
          description: Include all tickets (overrides column filter)
          required: false
          schema:
            type: boolean
            default: false
        - name: done_limit
          in: query
          description: Limit the number of done column tickets returned
          required: false
          schema:
            type: integer
        - name: filter_blocked
          in: query
          description: Exclude blocked, closed, and human_feedback tickets (optimized query)
          required: false
          schema:
            type: boolean
        - name: include_closed
          in: query
          description: "Whether to include closed tickets (default: true)"
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
    post:
      summary: Create a new ticket
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /projects/{slug}/tickets/{ticketId}:
    get:
      summary: Get ticket details
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ticket details with comments, dependencies, and history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a ticket
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a ticket
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: role_id
          in: query
          description: Role ID of the actor performing the deletion
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Ticket deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                      ticket_id:
                        type: integer
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{slug}/tickets/{ticketId}/transitions:
    get:
      summary: Get allowed transitions for a ticket
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Allowed transitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionsListResponse'
  /projects/{slug}/tickets/{ticketId}/move:
    post:
      summary: Move a ticket to another column
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTicketRequest'
      responses:
        '200':
          description: Ticket moved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      moved:
                        type: boolean
                      ticket_id:
                        type: integer
                      to_column:
                        type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/WorkflowError'

  # ==========================================================================
  # TICKET COMMENTS
  # ==========================================================================
  /projects/{slug}/tickets/{ticketId}/comments:
    post:
      summary: Add a comment to a ticket
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
  /projects/{slug}/tickets/{ticketId}/comments/{commentId}:
    patch:
      summary: Update a comment
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
    delete:
      summary: Delete a comment
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: integer
      responses:
        '200':
          description: Comment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                      comment_id:
                        type: integer

  # ==========================================================================
  # TICKET DEPENDENCIES
  # ==========================================================================
  /projects/{slug}/tickets/{ticketId}/dependencies:
    get:
      summary: Get ticket dependencies grouped by type
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dependencies grouped by type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependenciesResponse'
    delete:
      summary: Remove a dependency (query parameters)
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: depends_on_id
          in: query
          required: true
          description: The dependency ticket ID
          schema:
            type: integer
        - name: relation_type
          in: query
          required: true
          description: The dependency relation type
          schema:
            type: string
            enum:
              - blocked_by
              - depends_on
              - related
      responses:
        '200':
          description: Dependency removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      removed:
                        type: boolean
                      ticket_id:
                        type: integer
                      depends_on_id:
                        type: integer
        '400':
          $ref: '#/components/responses/ValidationError'
  /projects/{slug}/tickets/{ticketId}/dependencies/{relationType}/{depId}:
    post:
      summary: Add a dependency (path-based)
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: relationType
          in: path
          required: true
          description: The dependency relation type
          schema:
            type: string
            enum:
              - blocked_by
              - depends_on
              - related
        - name: depId
          in: path
          required: true
          description: The dependency ticket ID
          schema:
            type: integer
      responses:
        '201':
          description: Dependency added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      added:
                        type: boolean
                      ticket_id:
                        type: integer
                      depends_on_id:
                        type: integer
                      relation_type:
                        type: string
    delete:
      summary: Remove a dependency (path-based)
      tags:
        - Tickets
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: relationType
          in: path
          required: true
          schema:
            type: string
            enum:
              - blocked_by
              - depends_on
              - related
        - name: depId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dependency removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      removed:
                        type: boolean
                      ticket_id:
                        type: integer
                      depends_on_id:
                        type: integer

  # ==========================================================================
  # CONVERSATIONS
  # ==========================================================================
...
components:
  schemas:
...
    CreateTicketRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          default: ""
        column:
          type: string
          description: Target column slug
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 2
        labels:
          type: array
          items:
            type: string
        estimate:
          type: number
          nullable: true
        parent_id:
          type: integer
          nullable: true
          description: Parent ticket ID for sub-tasks
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 5
        labels:
          type: array
          items:
            type: string
        estimate:
          type: number
          nullable: true
        parent_id:
          type: integer
          nullable: true

    # ----- Comment -----
...
    MoveTicketRequest:
      type: object
      required:
        - to_column
      properties:
        to_column:
          type: string
          description: Target column slug
        comment:
          type: string
          maxLength: 5000
        role_id:
          type: integer

    # ----- Dependency -----
...
    TicketListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Ticket'
            total:
              type: integer
              description: Total number of matching tickets
            done_total:
              type: integer
              description: Total number of done tickets
    TicketResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Ticket'
    TicketDetailResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                project_id:
                  type: integer
                parent_id:
                  type: integer
                  nullable: true
                column:
                  type: string
                column_id:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                labels:
                  type: array
                  items:
                    type: string
                priority:
                  type: integer
                  minimum: 1
                  maximum: 5
                estimate:
                  type: number
                  nullable: true
                is_blocked:
                  type: boolean
                blocking_ticket_ids:
                  type: array
                  items:
                    type: integer
                comments:
                  type: array
                  items:
                    $ref: '#/components/schemas/Comment'
                status_history:
                  type: array
                  items:
                    $ref: '#/components/schemas/StatusHistoryEntry'
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                closed_at:
                  type: string
                  format: date-time
                  nullable: true
    TransitionsListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/TicketTransition'
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
...
    ValidationError:
      description: Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
...